pyproject.toml
[project]
name = "faas-demo"
version = "0.1.0"
description = "Function as a Service demo using Azure Functions"
authors = [
    {name = "Rishabh Chakrabarti", email = "rishabh.chakrabarti@happiestminds.com"},
]
dependencies = [
    "langchain>=0.0.329",
    "azure-functions",
    "fastapi==0.68.0",
    "uvicorn==0.15.0",
    "sqlalchemy==1.4.26",
    "jose==1.0.0",
    "python-jose==3.3.0",
    "passlib==1.7.4",
    "pydantic>=1.6.2,<2.0.0",
    "python-dotenv",
    "python-multipart",
    "bcrypt",
    "openai==0.28",
    "azureopenai",
]

requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
